refactor metric/retrieval_contents.py to p_pgt.py
