+++ autorag/evaluation/metric/__init__.py
@@ -1,4 +1,9 @@
 from .a_agt import bleu, meteor, rouge, sem_score, g_eval, bert_score
+from .p_pgt import (
+	retrieval_token_f1,
+	retrieval_token_precision,
+	retrieval_token_recall,
+)
 from .pid_pidgt import (
 	retrieval_f1,
 	retrieval_recall,
@@ -6,9 +11,4 @@ from .pid_pidgt import (
 	retrieval_mrr,
 	retrieval_ndcg,
 	retrieval_map,
-)
-from .retrieval_contents import (
-	retrieval_token_f1,
-	retrieval_token_precision,
-	retrieval_token_recall,
 )
